feat: Optimize database schema and enable compression
This commit optimizes the database schema for space efficiency and enables PostgreSQL's LZ4 compression.

- Tightened data types across all schema files, using more specific types like `REAL`, `DATE`, and `VARCHAR(n)` instead of `DOUBLE PRECISION` and `TEXT`.
- Enabled `lz4` compression in the `docker-compose.yml` for the PostgreSQL service to reduce storage for TOAST-able data.
- Addressed data integrity issues by reverting specific columns to `BIGINT` and `DOUBLE PRECISION` where necessary to prevent data overflow and precision loss.

+++ db/12-create-prices-table.sql
@@ -7,13 +7,13 @@ CREATE TABLE IF NOT EXISTS prices (
   "typeId" INTEGER NOT NULL,
   "regionId" INTEGER NOT NULL,
   "priceDate" DATE NOT NULL,
-  "averagePrice" DOUBLE PRECISION,
-  "highestPrice" DOUBLE PRECISION,
-  "lowestPrice" DOUBLE PRECISION,
+  "averagePrice" REAL,
+  "highestPrice" REAL,
+  "lowestPrice" REAL,
   "orderCount" INTEGER,
   "volume" BIGINT,
   "updatedAt" TIMESTAMP DEFAULT NOW(),
-  "version" BIGINT,
+  "version" INTEGER,
   PRIMARY KEY ("typeId", "regionId", "priceDate")
 );
